Выполнено ДЗ c дополнениями
